+++ lifelines/fitters/__init__.py
@@ -1130,7 +1130,8 @@ class ParametricRegressionFitter(RegressionFitter):
     _scipy_fit_options = dict()
     _KNOWN_MODEL = False
 
-    def __init__(self, alpha=0.05, penalizer=0.0, l1_ratio=0):
+    def __init__(self, alpha=0.05, penalizer=0.0, l1_ratio=0.0):
+
         super(ParametricRegressionFitter, self).__init__(alpha=alpha)
         self.penalizer = penalizer
         self.l1_ratio = l1_ratio
@@ -1644,7 +1645,7 @@ class ParametricRegressionFitter(RegressionFitter):
 
     def _add_penalty(self, params, neg_ll):
         params, _ = flatten(params)
-        # remove constant cols from being penalized
+        # remove intercepts from being penalized
         params = params[~self._constant_cols]
         if self.penalizer > 0 and self.l1_ratio > 0:
             penalty = self.l1_ratio * anp.abs(params).sum() + 0.5 * (1.0 - self.l1_ratio) * (params ** 2).sum()
@@ -2826,18 +2827,6 @@ class ParametericAFTRegressionFitter(ParametricRegressionFitter):
             if constant_col in mapping:
                 d[param][mapping.index(constant_col)] = _transform_ith_param(getattr(uni_model, param))
         return d
-
-    def _add_penalty(self, params, neg_ll):
-        params, _ = flatten(params)
-        # remove intercepts from being penalized
-        params = params[~self._constant_cols]
-        if self.penalizer > 0 and self.l1_ratio > 0:
-            penalty = self.l1_ratio * anp.abs(params).sum() + 0.5 * (1.0 - self.l1_ratio) * (params ** 2).sum()
-        elif self.penalizer > 0 and self.l1_ratio <= 0:
-            penalty = 0.5 * (params ** 2).sum()
-        else:
-            penalty = 0
-        return neg_ll + self.penalizer * penalty
 
     def plot(self, columns=None, parameter=None, ax=None, **errorbar_kwargs):
         """