+++ lifelines/plotting.py
@@ -1,10 +1,14 @@
 # -*- coding: utf-8 -*-
 
 import warnings
+from typing import Union
+
 import numpy as np
-from lifelines.utils import coalesce, CensoringType
 from scipy import stats
 from matplotlib import pyplot as plt
+import pandas as pd
+
+from lifelines.utils import coalesce, CensoringType
 
 
 __all__ = ["add_at_risk_counts", "plot_lifetimes", "qq_plot", "cdf_plot", "rmst_plot", "loglogs_plot"]
@@ -701,7 +705,7 @@ def _plot_estimate(
 
 
 class PlotEstimateConfig:
-    def __init__(self, cls, estimate, loc, iloc, show_censors, censor_styles, ax, **kwargs):
+    def __init__(self, cls, estimate: Union[str, pd.DataFrame], loc, iloc, show_censors, censor_styles, ax, **kwargs):
 
         self.censor_styles = coalesce(censor_styles, {})